+++ myceli/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "block-streamer"
-version = "0.1.0"
+name = "myceli"
+version = "0.5.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
